Preparing Unit/UnitEmitter for Local ArrayTypeTable
Summary:
This is part of a larger diff that prepares `Unit`/`UnitEmitter` for local `ArrayTypeTable` changes. The plan is to replicate what has been done to `LitstrTable`, where there is a global singleton for repo mode, and unit local instances for non-repo mode. Global litstr records have ID's beyond certain offset.

This specific diff contains repo scheme changes in the load/save unit/emitter logic. Lookup logic is yet to be modified to include the difference of global/local ids.

Reviewed By: markw65, mxw

Differential Revision: D5498920

fbshipit-source-id: eb95c590bc96fdca743f47c1045a29216d5e2c93