Resolve RepoAuthType in PreClassEmitter and FuncEmitter in Deserialization Time
Summary:
Depends on D5550924. This diff resolve `RepoAuthType::Array` at deserialization time in a couple key places. The resolution is needed since `RepoAuthType::Array` is serialized as just id's and when it is deserialized, the actual `Array*` may not be immediately decode-able in non-repo mode, since we need the unit local array type table. The key places of such resolution should happen in key deserialization places, such as `PreClassEmitter` and `FuncEmitter`. This is because, HHBBC is the one responsible for `RepoAuthType`, and HHBBC is producing `UnitEmitter`, which contains `PreClassEmitter` and `FuncEmitter`. All funcs, classes, are emitted from these emitters. As long as we can handle deserialization of these emitters correctly, we will be guaranteed to have fully resolved `RepoAuthType::Array`.

Other places where `array()` is called just for comparison with `nullptr` is replaced with other suitable test that do not rely on `RepoAuthType::Array` being resolved.

Reviewed By: mxw

Differential Revision: D5551769

fbshipit-source-id: 30dc57f748879d0dbea5a7bda40e54a8ade41e69