Move ArrayTypeTable from Repo::GlobalData to GlobalArrayTypeTable
Summary:
Depends on D5551769. This diff actually moves the array type table out of `Repo::GlobalData` and into `GlobalArrayTypeTable`. This makes array type table behaves the same as `LitstrTable`.

The serialization/deserialization of `GlobalArrayTypeTable` is done together with the ser/deser of `Repo::GlobalData`, to ensure that the code change behaves like before.

Reviewed By: mxw

Differential Revision: D5552097

fbshipit-source-id: b5a34a6be80e7c5243787294251a05f19d4d6fc0