Fix bytcode-as-text output for hhvm --hphp
Summary:
This was effectively broken by the lineTable caching we do. For every
bytecode, we would fetch the line-number, which involved fetching the
whole lineTable from the repo, getting the line number for the current
pc, caching it, and then throwing the line table away. The cached
value would obviously never be reused.

Pre-caching the lineTable (which was already possible), and actually
using it (which was not) speeds up the output by a factor of 100 or so.

At the same time, fix it to output the post-hhbbc bytecode, since
thats what we're interested in (and you can always get the pre-hhbbc
bytecode by turning hhbbc off).

This also sets things up to remove the locking in
LitstrTable::mergeLitstr, but I'll do that in a follow up diff.

Reviewed By: ricklavoie

Differential Revision: D5668519

fbshipit-source-id: ba3e2317bfcde49281152ca76a9d1daaf486cb19