Get rid of the mutex in LitstrTable
Summary:
The original plan for LitstrTable was that it would be in read-only
mode most of the time, but write-only mode while building the repo.

That was broken by two things: bytecode-to-text output interleaved
generating units with dumping units, which meant we had to keep
switching back and forth, and keep the litstr-table up to date; and we
(incorrectly) allowed systemlib units that hadn't been included at
build time to be used at runtime. Both those issues are now fixed, so
this cleans things up, and simplifies the code.

While doing this, I noticed that hhbbc would read in a LitstrTable at
the beginning, and then just update it at the end (which explains why
D5568720 was such a win - almost every symbol was already in the
table). That could result in no-longer-used strings being in the
LitstrTable. Fix that by killing the LitstrTable before creating the
new set of UnitEmitters.

Reviewed By: ricklavoie

Differential Revision: D5681456

fbshipit-source-id: 2f4eb551408d1d98ede5b2f5a2af94cc6b9d1678