Don't generate UnitSourceLoc entries when HHBBC is run stand-alone
Summary:
We don't need UnitSourceLoc entries in the repo for repo-auth mode because
they're only useful in the debugger. However, when HHBBC was run stand-alone (as
opposed to part of HPHPc), the RepoDebugInfo runtime option was left to its
default (true), so it was being emitted into the repo. This causes repos
generated such a way to be much larger than normal. Manually set the runtime
option to false to match HPHPc so that HHBBC will not generate UnitSourceLoc
entries.

Reviewed By: markw65

Differential Revision: D5686567

fbshipit-source-id: 0e30249489f57a39d32efb0587fec16433b42403