Initial plumbing for one-bit reference counting
Summary:
This diff adds a new build mode for one-bit refcounting, along with
enough of an implementation to make some things work in the
interpreter. RefCount is now an enum with an underlying type that changes based
on the presence of ONE_BIT_REFCOUNT. This makes it harder to accidentally write
a bogus reference count, but it obviously does nothing for jitted code. This
will come in later diffs.

The size and layout of HeapObject is unchanged, mostly because doing so would
upset the alignment of everything else. In ONE_BIT_REFCOUNT mode, the refcount
is 1 byte, followed by 3 bytes of padding. It's a full byte rather than a
single bit because we still need to support Static and Uncounted values. This
doesn't affect any of the reference counting operations or reduce any of the
benefits of this kind of reference counting, since the only predicate that
matters is still (m_count == 0).

Most of the changes are one of two things:
- Using InitialValue instead of 1 as the starting refcount for new
  objects. This is conditionally defined to 0 or 1, depending on the build
  mode.
- Using one of the new HeapObject::initHeader* variants. I decided it was worth
  being more explicit about which aux fields were used by which versions,
  rather than having a few different overloads that take types that all convert
  to each other. Object types that use m_aux32 now explicitly use m_aux32
  rather than sometimes using m_count, since m_count isn't always 32 bits.

EZC is completely unsupported, and will probably stay that way
indefinitely. All of the RefData::z* functions will fail in the new build mode.

Nothing in test/quick -m interp crashes in one-bit mode, but plenty of tests
fail because trying to instantiate an object with a destruct now fatals.

Reviewed By: edwinsmith

Differential Revision: D5661420

fbshipit-source-id: 19f9d7a326edb1325f81e34125d039e497e46496