A shape A with known fields may only be a supertype of another shape B if A contains every single field that B may contain
Summary:
Consider a shape A and shape B. In this example, we're trying to use shape A as a supertype of shape B.

Prior to this diff, we only verified that supertype shape A's fields were //consistent// with subtype shape B's fields. So, consider:

```
type A = shape('int_field' => int);
type B = shape('int_field' => int, ?'optional_string_field' => string);
```

Here, A has a single field. It is in fact consistent with B's definition! However, **since A is a `FieldsFullyKnown` shape**, it has the additional requirement that **no other fields may be present in the subtype shape B**. Here, `'optional_string_field'` may be present in B, and so A should not be a supertype of B.

The way that I've implemented this in the coe is as follows:

If A is `FieldsFullyKnown`, then we iterate over its fields, and ensure that each is defined in B. Later in the code, we ensure that the type are consistent; we merely need to make sure that A //declares// all of the possibly-present fields in B.

Note that we //already//, prior to this diff, enforce that B must be a `FieldsFullyKnown` shape.

Reviewed By: dlreeves

Differential Revision: D5680742

fbshipit-source-id: efab991fc65e5ba94a1febe4bc79dbca19be5a08