Dedup comment lowering effort
Summary: The lowerer had its own comment scouring facility to find line- & block comments and to grab `HH_FIXME`s and `HH_IGNORE_ERROR`s. In the interim of that being written, the full fidelity parser was extended to process trivia in more detail. It's now just redundant work. This diff uses the structural representation of the trivia, instead of reparsing. It's better perf, better maintainability and better consistency.

Reviewed By: vladima

Differential Revision: D5536844

fbshipit-source-id: 9df5d9df9f56886f1c46d29fca412431881854f2