Fix a couple of silence operator issues
Summary:
If a fatal was thrown from within a silenced region, and the
error_handler was called, there was nothing to reset the silence
state, and code running in the error_handler would be silenced.

If the verifier encountered a Fatal opcode with any silencers active,
it would fail verification. Instead, we should just clear the silencer
state (reflecting the change in the runtime to do just that).

Reviewed By: alexeyt

Differential Revision: D5693809

fbshipit-source-id: ea25abecad69b188287cdd27e3733ccc580bfc93