Eliminate conservative scanning of certian heap objects
Summary:
To enable two-phase scanning, we need to avoid conservative
scanning heap objects which aren't guaranteed to be reachable
from conservative-scanned roots (which is most heap objects).

Many of the pointers we were conservative-scanning are actually
known-pointers to unknown typed things. We can still scan these
exactly with a custom scanner.

Loosened the restrictions on the overloaded Scanner::scan() functions
by allowing "uninteresting" types; when we call these by hand, we
want to allow uninteresting types; the important semantics are that
we're scanning a pointer (vs some unknown-layout hunk of memory).

Removed type_scan::enqAddr() since it was dead.

Differential Revision: D5630732

fbshipit-source-id: eeef70e825cbf763655dea7536545855da4577e2