Add PopL bytecode
Summary:
Add a new bytecode called PopL. SetL does not pop the value it stores into the
local, but its extremely common to not need the value afterwards, so its usually
immediately followed by a PopC. PopL has the same semantics as a SetL/PopC pair,
but it might be implemented more efficiently.

When the local we're storing to isn't a ref, we can teleport the value directly
from the stack into the local without any ref-count manipulation. If the local
is a ref, we must still manipulate the ref-counts the same as SetL followed by
PopC. This is because eliding the ref-counts can change when destructors run if
the ref is also reachable via another destructor.

I did not modify the emitter to emit any PopL instructions directly. While its
easy in a few cases, in other cases its not so trivial, and I did not want to do
it halfway because then its difficult to keep HackC in sync. Instead I rely on
HHBBC doing a peephole optimization to replace such pairs.

I do not expect this to be much of a CPU win, because we should be optimizing
away SetL/PopC pairs into the same end result anyways, but this is semantically
cleaner and makes the bytecode smaller. It might also help profiling
translations a bit because we don't do the necessary ref-count optimizations
there.

Reviewed By: swtaarrs

Differential Revision: D5668382

fbshipit-source-id: 5ad42cf18820dbc947c240fb40aa1558f813b978