Handle passing parameters by unknown refiness better in HHBBC's minstr handling
Summary:
In HHBBC, when passing a parameter by unknown refiness, we need to compute its
effects both ways (as if it was taken by ref, and as if it wasn't). If the
parameter being passed is from a member instruction sequence, this means
handling the sequence as both defining and non-defining. Right now, we handle
this by interpreting each step in the sequence as both defining and
non-defining, merging the results, and then proceeding to the next step. In
fact, for the member instruction state, we do not merge at all, and simply
assert they are equivalent.

However, these assertions are too zealous. The member instruction state for the
non-defining and defining cases can diverge quite a lot. For instance, consider
a nullable array. In the non-defining case, when we index into the nullable
array, we no longer know what the base location is. If the array was null, the
base is now the immutable null base. If it was not, our base is inside the
array. In the defining case, however, the null will promote into an empty array,
so when we index into it, we always know the base is inside the array. So, after
the index step, we have two very different member instruction state, and its
difficult to reconcile the two. Right now we just assert.

A better way to deal with this is to instead track two separate member
instruction states for the define and non-define cases. Instead of trying to
unify them at each step, they're tracked separately, and the defining one will
ultimately affect the state of locals, properties, etc. To facilitate this, wrap
up the member instruction state into its own structure (MInstrState), and add a
second instance to State. When we're in a member instruction sequence of unknown
define-ness, we'll interp each step twice, once with the non-define state
"active", and once with the define state "active". To avoid overhead for this
rare case, the second member instruction state is heap allocated on demand with
a copy_ptr.

This diff has a bit of churn because of encapsulating the member instruction
state into a new struct, but I feel its worthwhile.

Reviewed By: markw65

Differential Revision: D5696672

fbshipit-source-id: 63f6dbd137a2c5fc554b6cac9aaef1600c675f0a