Add timeout to HPHPc and HHBBC invocations inside test runner
Summary:
Sometimes when working on HHBBC, you might trigger an infinite loop or
other slow behavior. When this happens, the test runner just hangs forever
because, unlike when running the actual test, we do not impose a timeout. Fix
the test runner to impose the same timeout as we do on the main test run.

Reviewed By: swtaarrs

Differential Revision: D5700031

fbshipit-source-id: cd87c1338293cce178f3a48a0992b89f625e2291