Fix optimization of static locals
Summary:
I broke this when I refactored it a while ago. If a
StaticLocCheck/StaticLocDef pair always sets the same, known value,
and the local isn't otherwise modified, we should be able to get rid
of the static local altogether; but it wasn't happening because we
didn't check for a known value until after setLocRaw killed it.

Reviewed By: ricklavoie

Differential Revision: D5695977

fbshipit-source-id: 842d638b96a0d67e94dacee01b020116465aade4