Remove ratToAssertType, use typeFromRAT directly
Summary:
When turning a RAT type into an IR type assertion, we use ratToAssertType() to
turn the RAT type into an appropriate JIT type. ratToAssertType() passes some
RAT types off to typeFromRAT(), but for others (mainly optional types), it does
not produce a type at all (which means no assertion is emitted). There's a vague
comment that this is necessary because asserting on some of the types regresses
performance, but I suspect this comment is outdated. I can't find any evidence
this is necessary. Get rid of ratToAssertType() and instead just use
typeFromRAT() directly.

Reviewed By: swtaarrs

Differential Revision: D5702891

fbshipit-source-id: aa2ab37f11661fbe93a012d9520e680972d4ac59