Allow optional specialized array types in RepoAuthTypes
Summary:
HHBBC was dropping specialized array types if the array was optional. Likewise,
the JIT was ignoring the specialized array information if the array was optional
(which never actually happens because HHBBC doesn't encode it). There doesn't
seem to be any reason to disallow this (both HHBBC and the JIT can deal with it
without issue), so allow it.

Reviewed By: mxw

Differential Revision: D5705506

fbshipit-source-id: a868f8346468656f4fa1d1f276e653e6d5c25079