Allow indexing into nullable shapes when part of null coalesce
Summary:
**Before**

```
function f(?MyShape $shape): void {
  $shape['optional_field'] ?? 12;
}
```
Produced the following typecheck error:

```
line 2, characters 3-26:
You are trying to access an element of this container but the container could be null.  (Typing[4063])
line 1, characters 12-19:
This is what makes me believe it can be null
```

**After**

We thread the indicator that we are on the left hand side of the null coalesce operator through the typechecker to add context when typechecking array_get for shapes. This allows indexing into nullable shapes so long as it is on the left hand side of the null coalesce.

```
  function f(?MyShape $shape): void {
    $shape['optional_field'] ?? 12;
  }
```

No longer produces an error.

The following still produce errors as intended:

```
function foo(): void {
  $bar = (true ? array('a' => 5) : null)['a'];
}
```

This will produce

```
File "nullable_shape_disallows_indexing.php", line 4, characters 10-45:
You are trying to access an element of this container but the container could be null.  (Typing[4063])
File "nullable_shape_disallows_indexing.php", line 4, characters 36-39:
This is what makes me believe it can be null
```

However, this change now allows the following past the typechecker:

```
function foo(): void {
  null['a'] ?? 12;
}
```

While this, on the face of it, may be undesirable, hhvm does allow this.

Reviewed By: michaeltingley

Differential Revision: D5690401

fbshipit-source-id: f176c27cf7255362338b8267e85c2a3f7a154c08