Emit doc comments for class properties
Summary:
This diff is a very rough implementation of doc comments for class properties. Currently we only emit doc comments for class properties if they are formatted as such
```
class C {
  /** hi */
  public int $x;
  /** hi */ private int $y;
}
```
We do not emit doc comments if the doc comment is in the middle of the property or after, including multiple declarations on the same line. The following types are not emitted correctly:
```
class C {
  public /** hi */ int $x;
  private int /** hi */ $a;
  private int /** hi */ $y, int /** hi */ $y;
}
```
More over, PHP/HHVM has this strange behavior where if a property has two doc comments, it drop all but last one. We also do not emit this correctly.
```
class C {
  public /** hi */ int $x;
  private int /** hi */ $a;
  private int /** hi */ $y /** lol */;
}
```

Reviewed By: hubyrod

Differential Revision: D5702079

fbshipit-source-id: f05f6fb2d949fa7a790e8a806892bf8fdf29174f