Skip loading of RAT-table in hhbbc startup
Summary:
When running standalone, hhbbc loads the globalData in order to set up
runtime options before calling process init. Since we only need the
runtime options, its pointless to load the RAT-array table - but its
also not safe, because we haven't initialized the runtime yet.

Normally this isn't an issue, because the table created by hphpc is
empty, but if we try to run hhbbc on a repo thats already been
processed by hhbbc, it asserts.

This adds an option to skip the table, and uses it for this one call.

Reviewed By: ricklavoie

Differential Revision: D5715486

fbshipit-source-id: a7c38658faa119aa6aa898db9b8e9f0b32f555a6