Fix offset counting in multiline delimited comments
Summary:
We have an off-by-one error here, since normally the newline character should be counted (via Ignore) before the Newline split, but in multiline delimited comments, we count the character after the split.

The symptom of this problem is that when range-formatting begins inside a delimited comment, we incorrectly delete a newline at the beginning of the range, joining the first line in the range to the line preceding it.

Reviewed By: arxanas

Differential Revision: D5710818

fbshipit-source-id: 99ac531770b435ea036690ea963ef2653af5799c