Improve type widening in HHBBC
Summary:
In HHBBC, types do not necessarily have finite height under unioning. That is,
you can union together an arbitrary number of types and never arrive at a fixed
point. This obviously can prevent termination, so HHBBC deals with it with a
heuristic. If we iterate too much without reaching a fixed point, we start to
use a "widening union", which *is* guaranteed to eventually reach a fixed point.

However, the widening union is implemented quite bluntly right now. We just drop
any extra data with the type (for example, array specializations). As a result,
our heuristics have to be quite conservative since widening drops so much
information. Widening doesn't really happen in normal code, at least right
now. Experimentations with better array specializations have caused it to
trigger more in real code, so its worthwhile to improve this handling.

Instead of dropping any extra data from the type, the widening operation now
bounds the depth of the type. If there's a leaf of the type which exceeds the
max depth, the type at that max depth has its specialized information dropped
(not the entire type). Since the type can never grow larger than the max depth,
it is finite and will eventually reach a fixed point under unioning. In theory,
there can be a similar issue with a type's width, but in practice right now a
type's width cannot grow arbitrary large.

Since this new widening is not as drastic as before, we can be more aggressive
with our heuristics. In addition, we failed to apply the widening heuristics to
static local types, which meant it was easy to drive HHBBC into an infinite
loop.

Reviewed By: markw65

Differential Revision: D5702036

fbshipit-source-id: f1efbcef362265c5242b36a46bae663629705c05