Print trailing newlines in State_printer
Summary:
Previously, when printing a solution for a chunk group, State_printer produced a string with a leading newline and no trailing newline. This was convenient to implement, since chunks are associated with the split preceding them. However, we rarely actually want a formatted string with a leading newline and no trailing newline. More frequently, we are formatting a range of lines (in which case, the range omits the leading newline and includes the trailing newline) or we are formatting a single node (in which case, I think it's acceptable to include a trailing newline when the last token would have a trailing newline in the formatted file).

This change modifies State_printer to not print a newline before the first chunk in a group, and to print a newline after the last chunk in the group. If the last chunk in the printed range is not the last chunk in the group, a trailing newline is printed if the chunk following the last chunk in the range is bound to have a newline before it in the solve state.

Differential Revision: D5635262

fbshipit-source-id: 320ec73359256256461448bec1f955d6373af08c