Refactor state printing to behave predictably at non-line-boundaries
Summary:
Currently, Libhackfmt.format_range makes no promises about the behavior of whitespace at the beginning and end of a formatting range which does not fall at line boundaries, since the range formatting implementation was originally written for formatting at line boundaries, and non-line-boundary formatting was hacked on later. This change chooses semantics for format_range that are reasonable for both formatting at line boundaries and formatting arbitrary ranges of tokens, and refactors Chunk and State_printer to implement them.

The first major change this diff makes is to change the way text is stored in chunks. Before this change, chunks maintained a string of the text they represented, and as more atoms were added, they were concatenated onto this string. An `atoms` list maintained a mapping between each atom's offset in the original source and its offset in the chunk's `text` string. After this change, chunks don't maintain a `text` string at all--the text a chunk represents is the concatenation of the text of its atoms.

The second major change is to refactor State_printer so that its behavior is more understandable. Before this change, State_printer attempted to write directly into an output buffer while iterating over chunks, in a way which was error-prone and not correct or consistent at non-line-boundaries. This change introduces the concept of `subchunks`, which are either atoms, trailing commas, or whitespace. An atom subchunk is associated with its offset in the original source; other subchunk variants are not. When range formatting, we filter to the atoms which were present in the requested range (including their leading and trailing whitespace/commas), then chop off leading newlines and trailing spaces. This makes range-formatting behavior much more predictable, while still being suitable for formatting at line boundaries and as-you-type formatting.

Reviewed By: arxanas

Differential Revision: D5702779

fbshipit-source-id: 2ad616067cd82a88943814a22138964370a11559