Parallelize repo-writing with whole_program
Summary:
Writing the repo is one of the slowest steps, because we can't
effectively parallelize it. Prior to adding hhbbc, we found that the
fastest approach was to use multiple threads to generate the data, but
a single thread to write it. That code is still present, but we only
use it when hhbbc is turned off. In hhbbc mode, we generate all the
units, and once we've finished, we start writing them to the repo.

This diff re-uses the old approach to start writing the repo as soon
as we have units available.

Reviewed By: ricklavoie

Differential Revision: D5720531

fbshipit-source-id: 1750cf9d273bf0ed641b775ee1f30c496e83903b