Implement a coroutines debugger test
Summary:
Inspired by hphp's [debugger tests](hphp/test/quick/debugger/), this diff extends run_hhvm.py to support a `--mode debug` option. Under this mode, instead of executing the code, the test driver will attach a debugger and consume an input source providing the debugging interaction. If the test is name `path/to/test.php`, then the test driver will look for a file `path/to/test.debug_in`. This file should be a sequence of key presses entered //exactly// as if the user was part of an hphpd session debugging the original source file (`path/to/test.php`). The debugger is automatically terminated at the end of the input, if `quit` is not manually invoked in the test. The test driver produces an abbreviated output from the debugger into an expectation file for verification.

If a test file //does not// have a corresponding .debug_in file, then the debug test is skipped for that test.

**Overall design:** https://fburl.com/coroutine-debugging

Reviewed By: vladima, ericlippert

Differential Revision: D5616336

fbshipit-source-id: a9a80c62f7355429b4723555d9b44365d86a4db9