preserve evaluation order
Summary:
This is a candidate version of preserving evaluation order for coroutines.
The crux of the idea - during a post-order rewrite pass we accumulate some extra information about nodes, specifically:
- function calls in `suspend` operator are replaced with member access `$closure->coroutineResultData` that will store the result of the call to coroutine. As extra information we store a prefix - list of statements that should be executed to initialize `$closure->coroutineResultData`
- for non-leaf nodes we can grab all prefixes that was accumulated for its children and recombine it into a prefix for a current node, i.e for conditional operators (`test`, `test_prefix`, `consequence`, `consequence_prefix`, `alternative`, `alternative_prefix`) the entire node will be replaced with freshly allocated `$closure->coroutineResultData` and final prefix will be

  ...test_prefix;
  if (test()) {
      ...consequence_prefix
      $closure->coroutineResultData = consequence;
  }
  else {
      ...alternative_prefix;
      $closure->coroutineResultData = alternative;
  }

In general with the exception of short circuiting operators if node has a list of child nodes: `c1, c2 c3, c4`
and if `c3` is the first node from the right that has prefix (meaning it has suspend somewhere in it) then all nodes on its left should be spilled (for each node spilling code goes to the prefix of the node) and replaced with a fresh `$closure->coroutineResultData` and then final prefix is the result of concatenation of all prefixes for child nodes.

- `&&`, `||`, `??`, `?:` operators are supported include suspends in tail positions

Reviewed By: michaeltingley, ericlippert

Differential Revision: D5565917

fbshipit-source-id: 3ac4339764599e2360a23ee330b3c965e222014d