Remove incorrect assertion
Summary:
The assertion is incorrect for Dim when the base is known to contain
the given element.

It would also be reasonable for various Base ops to be nothrow (eg a
BaseL where we know the local isn't uninit *is* nothrow - although
hhbbc doesn't currently mark it as such).

The test case is more complex than needed - I was constructing it to
test another diff I'm working on, but hit this bug first.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D5726268

fbshipit-source-id: b92370376c5a72076b1eb302b71b56e7c644705e