Make split between lambda arrow and XHPExpression parental
Summary: When an XHPExpression which spans across multiple lines is the body of a LambdaExpression, we would like to split before the XHPExpression and indent it, so that the opening tag is at the same indentation level as the closing tag. Making the split before the body Parental when the body is an XHPExpression ensures this behavior.

Reviewed By: arxanas

Differential Revision: D5727919

fbshipit-source-id: 38915132977e1a796d851732d7f511ffb215731b