Speed up the optimize loop
Summary:
hhbbc has an early constant-propagation phase, thats intended to turn
arrays constructed by AddElemC etc into static arrays, so that later
phases don't have to keep re-doing that work. If we turn off constant
propagation in hphpc, hhbbc's constant propagation phase takes a long
time. A big part of the problem is that we keep calling tv stack
values, so each time we call AddElemC we generate a new static array
with one more element than the last one (which is O(n^2) time and
space).

But it turns out that we're also spending a lot of time in
copy-on-write for the Types representing the arrays themselves, and I
just realized that thats because we save the entire state before each
instruction to use in the peephole phase, which forces the ArrLikeMap
to copy-on-write.

Peephole doesn't really need much of the state, so just give it what
it wants.

Reviewed By: ricklavoie

Differential Revision: D5733970

fbshipit-source-id: b76907baf5f5565789f8fb03326d30d2793b81eb