Leverage HH_logger.level to partially clean up integration_ml test output
Summary: Reconfigure methods in Hh_logger so that calls to log and log_duration go through the Level module at a default severity of Info. This way applications that wish to silence the output of the server test (ie, integration_ml tests) can simply set the level to OFF without affecting logging behavior in production.

Differential Revision: D5690336

fbshipit-source-id: bdb925cb7a1837e993f84ca87431963dffc157c2