First pass at a dependency analyzer on the AST
Summary:
This diff adds unit tests and a small library for calculating a dependency graph at the AST level.

It adds a few things:

- An argument to hh_single_type_check "--dump-toplevel-deps" (which currently doesn't do anything other than print out a few hint types)

- A unittest framework based on verify

- A new visitor on the AST which we'll use to calculate dependencies.

At the end of this diff stack, the goal is to be able to generate a toplevel-granularity simple dependency graph given only an AST and Global Naming information.

For now, we print out the results so that we can unit test some basic functionality, but I'll add an actual place to save the dependencies(probably shared memory) and dump them, just like hh_single_type_check --dump-deps does today.

Reviewed By: dabek

Differential Revision: D5654910

fbshipit-source-id: 21acaa1bbd73f3ecf4e386a6fb5bba43d7a692b1