Call tv less
Summary:
Two of the hottest calls were only checking to see if the type
represented a constant value, and didn't actually care what it
was. For ArrLikeMap and ArrLikePacked computing the value is
expensive, so I created is_scalar, which just determines whether
calling tv would return a cell.

But it turns out that in both those cases, if tv returned a value, it
would call tv again later in the same function, so I rewrote both of
those call sites to save the value from the first call, leaving no hot
uses of is_scalar. I'm planning more constant optimizations, however,
and I think it could be useful.

Reviewed By: ricklavoie

Differential Revision: D5736901

fbshipit-source-id: ab853c8006472b8a9e9435fb916a5c4e65fca32d