Fix the ReparseCompare test, and use it to fix inconsistencies in our generated syntax for coroutines
Summary:
= What is the `ReparseCompare` test?

You're probably unfamiliar with this test, so here's some background. Here's how it //should// work:

  1. Lower coroutines.
  2. Grab a snapshot of the resulting FFP parse tree.
  3. Print out the result to text.
  4. Reparse that text using the FFP (and no coroutine lowering this time).
  5. Compare the tree from step 4. with the tree from step 2., and print out any differences.

This would help us identify locations where we were //generating// code (from coroutines) in a way that was inconsistent with how the FFP would //parse// code.

= This change

Prior to this diff, the `ReparseCompare` test had a bug where it would reparse the tree //before// generating the "before" tree in the reparse-compare test. This meant that we were always comparing a "reparsed" tree against a "reparsed" tree, and so the test would erroneously produce no errors, even though there were some.

This diff corrects this problem in the coroutine test driver. (I also formatted the tree that can be obtained with `--syntax-tree` so that it is easier to interpret visually.) Consequently, this means that we had lots of test failures, because there are various small ways in which we generate code differently from how the FFP parses it. I've fixed those problems in this diff, as well.

= Why do we care?

In a future diff, we'll use the `FullFidelityAst` to lower out coroutines (currently, we do it manually). The `FullFidelityAst` has an expectation that the FFP parse tree is consistent with how the FFP would produce it, otherwise, it throws exceptions.

**Overall design:** https://fburl.com/coroutine-debugging

Reviewed By: ericlippert

Differential Revision: D5717984

fbshipit-source-id: fe4067aa794280831574ecab29ce53d82ac45b64