Trim empty `Markup` node from FFP result before comparing to legacy
Summary: Since we didn't parse markup as trivia anymore, we introduced a default empty `Markup` node at the start of every (lowered) AST. This broke `hh_single_parse`'s `--parser compare` by always reporting this difference. This diff dumps the first node of the AST if it is indeed an empty `Markup` node.

Reviewed By: ericlippert

Differential Revision: D5697525

fbshipit-source-id: e1dac6766aa0e7951f781f1330b7ea052ad9f34a