Added missing case for PHP "execution string" syntax in FFP lowerer
Summary: PHP has "eval everwhere" so why not in the middle of expressions. Because HackC must pass on HHVM's PHP tests, the FFP's lowerer shouldn't choke on the presense of these. There is no meaningful interpretation for them in the codegen at the moment, so stuff a `Null` in their position for now.

Reviewed By: hubyrod

Differential Revision: D5697585

fbshipit-source-id: cb9f7e345a95c235eb4ba2b6dc7166ab896e11ca