Allow function declarations everywhere in the full fidelity parser
Summary:
In PHP, any expression is statement, but also, any statement can be a declaration. The FFP must allow function declarations everwhere, e.g.
```
function foo() {
  function bar() {
    echo "clowntown";
  }
}
foo();
bar();
```
The Full Fidelity Parser already handles this case, but was missing the case for when these functions are `async`.

Simply adding the case for `async function` declarations surfaces the ambiguity when that is not the start of a declaration, but of an lambda-valued expression. Thus, the FFP must be forgiving when it gets tricked into believing there's a declaration, but it turns out to just be verbose lambda notation, e.g.
```
function foo() {
  async function () { return; };
}
```

Reviewed By: andrewjkennedy

Differential Revision: D5697612

fbshipit-source-id: 6446f332b1b8dbdf01665f22a51b3608977eab4d